moved auth verifier code to auth.rs
