+++ contracts/axelar-gateway/src/contract.rs
@@ -1,14 +1,13 @@
-use core::panic;
-use axelar_soroban_interfaces::types::{Message, ProofSignature, ProofSigner, WeightedSigner};
+use axelar_soroban_interfaces::types::{Message};
 use soroban_sdk::xdr::ToXdr;
-use soroban_sdk::{contract, contractimpl, crypto::Hash, panic_with_error, Address, Bytes, BytesN, Env, Vec, String};
+use soroban_sdk::{contract, contractimpl, panic_with_error, Address, Bytes, BytesN, Env, Vec, String};
 
 use crate::storage_types::{DataKey, MessageApprovalKey, MessageApprovalValue};
 use crate::types::CommandType;
-use crate::{error::Error, event};
+use crate::{error::Error, event, auth};
 use axelar_soroban_interfaces::{
     axelar_gateway::AxelarGatewayInterface,
-    types::{Proof, WeightedSigners},
+    types::{WeightedSigners},
 };
 
 #[contract]
@@ -37,29 +36,7 @@ impl AxelarGatewayInterface for AxelarGateway {
 
         env.storage().instance().set(&DataKey::Operator, &operator);
 
-        env.storage().instance().set(&DataKey::Epoch, &0_u64);
-
-        // TODO: Do we need to manually expose these in a query, or can it be read directly off of storage in Stellar?
-        env.storage().instance().set(
-            &DataKey::PreviousSignerRetention,
-            &previous_signer_retention,
-        );
-
-        env.storage()
-            .instance()
-            .set(&DataKey::DomainSeparator, &domain_separator);
-
-        env.storage()
-            .instance()
-            .set(&DataKey::MinimumRotationDelay, &minimum_rotation_delay);
-
-        if initial_signers.is_empty() {
-            panic_with_error!(env, Error::InvalidOperators);
-        }
-
-        for signers in initial_signers.into_iter() {
-            Self::rotate_signers_set(&env, signers.clone(), false);
-        }
+        auth::init_auth_verifier(env, domain_separator, minimum_rotation_delay, previous_signer_retention, initial_signers);
     }
 
     fn call_contract(
@@ -160,7 +137,7 @@ impl AxelarGatewayInterface for AxelarGateway {
             .keccak256(&(CommandType::ApproveMessages, messages.clone()).to_xdr(&env))
             .into();
 
-        Self::validate_proof(&env, data_hash.clone(), proof.clone());
+        auth::validate_proof(&env, data_hash.clone(), proof.clone());
 
         if messages.is_empty() {
             panic_with_error!(env, Error::EmptyMessages);
@@ -205,7 +182,7 @@ impl AxelarGatewayInterface for AxelarGateway {
             panic_with_error!(env, Error::RotationAlreadyExecuted);
         }
 
-        let is_latest_signers = Self::validate_proof(&env, data_hash.clone(), proof.clone());
+        let is_latest_signers = auth::validate_proof(&env, data_hash.clone(), proof.clone());
         if !is_latest_signers {
             panic_with_error!(env, Error::NotLatestSigners);
         }
@@ -214,7 +191,10 @@ impl AxelarGatewayInterface for AxelarGateway {
             .persistent()
             .set(&DataKey::RotationExecuted(data_hash), &true);
 
-        Self::rotate_signers_set(&env, signers, true);
+        let operator: Address = Self::operator(&env);
+        operator.require_auth();
+
+        auth::rotate_signers_set(&env, signers, true);
     }
 
     fn transfer_operatorship(env: Env, new_operator: Address) {
@@ -256,175 +236,6 @@ impl AxelarGateway {
 
     fn message_approval_hash(env: &Env, message: Message) -> MessageApprovalValue {
         MessageApprovalValue::Approved(env.crypto().keccak256(&message.to_xdr(env)).into())
-    }
-
-    fn validate_proof(env: &Env, data_hash: BytesN<32>, proof: Proof) -> bool {
-        let signer_set = proof.weighted_signers();
-
-        let signer_hash: BytesN<32> = env.crypto().keccak256(&signer_set.to_xdr(&env)).into();
-
-        let signer_epoch: u64 = env
-            .storage()
-            .persistent()
-            .get(&DataKey::EpochBySignerHash(signer_hash.clone()))
-            .unwrap_or(0);
-
-        if signer_epoch == 0 {
-            panic_with_error!(env, Error::InvalidSigners);
-        }
-
-        let current_epoch: u64 = Self::epoch(&env);
-
-        let is_latest_signers: bool = signer_epoch == current_epoch;
-
-        let previous_signer_retention: u64 = env
-            .storage()
-            .instance()
-            .get(&DataKey::PreviousSignerRetention)
-            .unwrap();
-
-        if current_epoch - signer_epoch > previous_signer_retention {
-            panic_with_error!(env, Error::InvalidSigners);
-        }
-
-        let msg_hash = Self::message_hash_to_sign(&env, signer_hash, data_hash);
-
-        if !Self::validate_signatures(&env, msg_hash, proof.clone()) {
-            panic_with_error!(env, Error::InvalidSignatures);
-        }
-
-        is_latest_signers
-    }
-
-    fn message_hash_to_sign(env: &Env, signer_hash: BytesN<32>, data_hash: BytesN<32>) -> Hash<32> {
-        let domain_separator: BytesN<32> = env
-            .storage()
-            .instance()
-            .get(&DataKey::DomainSeparator)
-            .unwrap();
-
-        let mut msg: Bytes = domain_separator.into();
-        msg.extend_from_array(&signer_hash.to_array());
-        msg.extend_from_array(&data_hash.to_array());
-
-        // TODO: use an appropriate non tx overlapping prefix
-        env.crypto().keccak256(&msg)
-    }
-
-    fn rotate_signers_set(env: &Env, new_signers: WeightedSigners, enforce_rotation_delay: bool) {
-        let operator: Address = Self::operator(&env);
-        operator.require_auth();
-
-        Self::validate_signers(env, &new_signers);
-
-        Self::update_rotation_timestamp(env, enforce_rotation_delay);
-
-        let new_signer_hash: BytesN<32> = env
-            .crypto()
-            .keccak256(&new_signers.clone().to_xdr(env))
-            .into();
-        let new_epoch: u64 = Self::epoch(&env) + 1;
-
-        env.storage().instance().set(&DataKey::Epoch, &new_epoch);
-
-        env.storage()
-            .persistent()
-            .set(&DataKey::SignerHashByEpoch(new_epoch), &new_signer_hash);
-
-        // If new_signers has been rotated to before, we will overwrite the epoch to point to the latest
-        env.storage().persistent().set(
-            &DataKey::EpochBySignerHash(new_signer_hash.clone()),
-            &new_epoch,
-        );
-
-        event::rotate_signers_set(env, new_epoch, new_signers, new_signer_hash);
-    }
-
-    fn update_rotation_timestamp(env: &Env, enforce_rotation_delay: bool) {
-        let minimum_rotation_delay: u64 = env
-            .storage()
-            .instance()
-            .get(&DataKey::MinimumRotationDelay)
-            .unwrap();
-
-        let last_rotation_timestamp: u64 = env
-            .storage()
-            .instance()
-            .get(&DataKey::LastRotationTimestamp)
-            .unwrap_or(0);
-
-        let current_timestamp = env.ledger().timestamp();
-
-        if enforce_rotation_delay
-            && (current_timestamp - last_rotation_timestamp < minimum_rotation_delay)
-        {
-            panic_with_error!(env, Error::InsufficientRotationDelay);
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::LastRotationTimestamp, &current_timestamp);
-    }
-
-    fn validate_signatures(env: &Env, msg_hash: Hash<32>, proof: Proof) -> bool {
-        let mut total_weight = 0u128;
-
-        for ProofSigner {
-            signer:
-                WeightedSigner {
-                    signer: public_key,
-                    weight,
-                },
-            signature,
-        } in proof.signers.iter()
-        {
-            if let ProofSignature::Signed(signature) = signature {
-                env.crypto()
-                    .ed25519_verify(&public_key, msg_hash.to_bytes().as_ref(), &signature);
-
-                total_weight = total_weight.checked_add(weight).unwrap();
-
-                if total_weight >= proof.threshold {
-                    return true;
-                }
-            }
-        }
-
-        false
-    }
-
-    /// Check if signer set is valid, i.e signer/pub key hash are in sorted order,
-    /// weights are non-zero and sum to at least threshold
-    fn validate_signers(env: &Env, weighted_signers: &WeightedSigners) {
-        if weighted_signers.signers.is_empty() {
-            panic_with_error!(env, Error::InvalidSigners);
-        }
-
-        // TODO: what's the min address/hash?
-        let mut previous_signer = BytesN::<32>::from_array(env, &[0; 32]);
-        let mut total_weight = 0u128;
-
-        for signer in weighted_signers.signers.iter() {
-            if previous_signer >= signer.signer {
-                panic_with_error!(env, Error::InvalidSigners);
-            }
-
-            if signer.weight == 0 {
-                panic_with_error!(env, Error::InvalidWeights);
-            }
-
-            previous_signer = signer.signer;
-            total_weight = total_weight.checked_add(signer.weight).unwrap();
-        }
-
-        let threshold = weighted_signers.threshold;
-        if threshold == 0 || total_weight < threshold {
-            panic_with_error!(env, Error::InvalidThreshold);
-        }
-    }
-
-    fn epoch(env: &Env) -> u64 {
-        env.storage().instance().get(&DataKey::Epoch).unwrap()
     }
 
 }
