+++ contracts/axelar-gateway/src/error.rs
@@ -9,13 +9,19 @@ pub enum Error {
     RotationAlreadyExecuted = 3,
     NotLatestSigners = 4,
     InvalidOperators = 5,
-    InvalidThreshold = 6,
-    DuplicateOperators = 7,
-    MalformedSigners = 8,
-    LowSignaturesWeight = 9,
-    InvalidProof = 10,
-    InvalidSigners = 11,
-    InsufficientRotationDelay = 12,
-    InvalidSignatures = 13,
-    InvalidWeights = 14,
+}
+
+#[contracterror]
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[repr(u32)]
+pub enum AuthError {
+    InvalidThreshold = 1,
+    DuplicateOperators = 2,
+    MalformedSigners = 3,
+    LowSignaturesWeight = 4,
+    InvalidProof = 5,
+    InvalidSigners = 6,
+    InsufficientRotationDelay = 7,
+    InvalidSignatures = 8,
+    InvalidWeights = 9,
 }
