+++ contracts/axelar-soroban-interfaces/src/axelar_gateway.rs
@@ -56,5 +56,4 @@ pub trait AxelarGatewayInterface {
     fn transfer_operatorship(env: Env, new_operator: Address);
 
     fn operator(env: &Env) -> Address;
-
 }
